v.7
troll

+++ Events/message.js
@@ -1,22 +1,39 @@
 const Discord = require("discord.js");
 const prefix = "&";
+let explosion = false ;
 
 module.exports = async(client, message) => {
     
     if(message.channel.type === "dm") return;
 
-    if(message.channel.name == "suggestions"  && ((message.content.startsWith(prefix) && message.content.slice(prefix.length).trim().split(/ +/g)[0] != "suggestion" ) || !message.member.roles.some(r=>["Dev", "Rainbow Kazekage","Scan Bot"].includes(r.name)))){
+    
+    if(message.channel.name == "suggestions"  && ((message.content.startsWith(prefix) && message.content.slice(prefix.length).trim().split(/ +/g)[0] != "suggestion" ) || !message.member.roles.cache.some(r=>["Dev", "Rainbow Kazekage"].includes(r.name)))){
         message.delete(); 
         return;
     }
+
+    if (explosion && !message.content.startsWith(prefix) && message.member.roles.cache.some(r=>["Rainbow Kazekage"].includes(r.name) )){
+        message.delete();
+        message.channel.send('bite');
+    }
     if(!message.content.startsWith(prefix)) return;
-    message.delete();
-    const args = message.content.slice(prefix.length).trim().split(/ +/g);
-    const commande = args.shift();
-
+    
+    const args = message.content.slice(prefix.length).trim().split(/ +/g); // .slice = enleve prefix ; .trim = enleve espace ; .split args[] chaque espace nv mot args
+    const commande = args.shift(); // mets nom commande dans commande et enleve première case args;
+    if(commande == "exploosion" ){
+        if(!message.member.roles.cache.some(r=>["Rainbow Kazekage"].includes(r.name) )){
+            explosion = !explosion;
+        }   
+        message.delete();   
+    }
     const cmd = client.commands.get(commande);
-
+    
     if(!cmd) return;
 
+    message.delete();
     cmd.run(client, message, args);
 };
+
+
+
+
